// src/favorites/common/enums/errors.enum.ts
export enum FAVORITES_ERRORS {
  INVALID_ID = 'Id is not a UUID',
  TRACK_NOT_EXIST = 'Track with this id does not exist',
  ALBUM_NOT_EXIST = 'Album with this id does not exist',
  ARTIST_NOT_EXIST = 'Artist with this id does not exist',
  NOT_FAVORITE_TRACK = 'Track is not favorite',
  NOT_FAVORITE_ALBUM = 'Album is not favorite',
  NOT_FAVORITE_ARTIST = 'Artist is not favorite',
}